booking-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.27.4
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0
